Make smoothing of TiledMapView customizable
Rendering of a tiled map view is hardcoded to be done with smoothing
enabled. Smoothing property is not accessible because it is set in an
internal object which is not reachable from outside. This change
makes the smoothing parameter for the drawVertices call configurable
on creation of the TiledMapView object.

An unnecessary "val" string was cleaned up in the parameter list of
TiledMapView, too.