Added virtual frame to BitmapSlice
The size and position/offset of the original untrimmed texture is stored
in the virtFrame property in BitmapSlice. The virtFrame is set up in the
Altas object with details of the region info. It will be used to
calculate anchorDispX and anchorDispY in the RectBase class when setting
up the vertices before they are drawn.

Unnecessary changes from the first commit were reverted.