Rework virtual Frame in BitmapSlice
The virtFrame property is not nullable in BitmapSlice. That means it is
null when the slice is not a trimmed texture. Offset and size of the
frame is stored in BmpSlice and calculated depending on trimmed state.

The properties anchorDispX and Y are now overriden in BaseImage instead
of RectBase. That should not break other code which is using RectBase
directly.