Add Layer, FrameTag and Slice config to AtlasInfo
In Aseprite it is possible to save in the sprite atlas config file
also information about layers, frame tags and slices. This commit
adds support for reading this additional information also into the
AtlasInfo object. If this information is not present in a json
file then the three lists "layers", "frameTags" and "slices" will
be empty.