Changed watcher function to use Dispatcher IO
The implementation of the watch functionality in the Jvm implementation
was changed to use Dispatcher IO instead of the executeIO function. That
was inspired by:

https://github.com/vishna/watchservice-ktx/blob/master/src/main/kotlin/dev/vishna/watchservice/watchservice.kt

Now it looks like that the watch function is working. But it is not
suspending any more the main thread. That means no launch is needed in
the test case.

Maybe later it would be an idea to use also coroutines.channels as API
to the application.