Remove invisible layers from image data of ASE file
Each cell of a frame is checked if it is located on an invisible (hidden)
layer. If yes, than that cell is dropped from the final image data. This
effectively means that no image data from any invisible layer is
available for drawing in the image data.

A new list of all visible layers is created and added to the image data.
The indexes of all visible layers are adjusted so that all entries have
the correct index. That is needed in ImageAnimationView class for
referencing the correct layer object. For that a change from "val index"
to "var index" in ImageLayer class was needed.

ASETest.kt was updated with a new ase file which contains
hidden/invisible layers. In order to test also a ase file with slices
the existing slice-example.ase was updated with a second invisible
layer. Both files are read correctly and the images are taken into the
generated atlas.