Fix slicing in ase files
The origin (x,y) of the rectangle which is used to cut the slice out of
the frame was wrong.

The reason is that the bitmap which should be sliced can be smaller as
the actual size of the frame. Aseprite saves only the visible pixels in
a bitmap. It does not include empty space around it. Thus the offset of
the bitmap needs to be taken into account. That means the offset
(targetX and targetY) needs to be substracted from the rectangle
(sliceFrame).

In other words values in "slice.bounds" are relative to the origin of
the bitmap while values in "sliceFrame" are relative to the origin of
the frame. TargetX and targetY is specifying the position of the bitmap
inside of the frame.