Refactor RepeatedImageView into SingleTile under tiles
Abstract base class "BaseTileMap" was created which contains everything
which is needed to draw tiles from a TileSet object. Then that
abstract class is implemented as TileMap and SingleTile. SingleTile
class is a specialization for a TileMap with only one tile. This can be
then used to implement the "repeated" Image View. The SingleTile class
implements the SmoothedBmpSlice interface which is needed for
integration into ImageAnimationView.

Changes were checked in example "tiled-background" that they do not
break the original TileMap implementation.