Add rigidbody and destruct component and update systems
Rigidbody and destruct components were added which are use to control
movement and destruction of entities. For that a new system called
DestructSystem was added. With it the destruction of an entity can
triger creation of new objects like explosions, etc.
Rigidbody component details will control how an entity is influenced
by gravity, friction or damping. This still needs to be implemented yet
in the MoveSystem.